requirements.txt
requests
python-dotenv
backports.zoneinfo; python_version<'3.9'
tzdata